Add answers system

+++ client/dialogs/editor.py
@@ -1,6 +1,7 @@
 from copy import deepcopy
+from typing import Union
 
-from PyQt5 import QtCore, QtWidgets
+from PyQt5 import QtCore, QtGui, QtWidgets
 
 import gui
 import models as m
@@ -10,19 +11,44 @@ from api import quiz as q
 
 class QuizEditor(QtWidgets.QDialog):
     ui: gui.dialogs.editor.Ui_QuizEditor
-    context: "widgets.quiz.Quiz"
-    quiz: m.Quiz
+    context: Union[QtWidgets.QWidget, "widgets.quiz.Quiz"]
+    token: str
+    quiz_id: int | None
+    quiz: q.Quiz
 
     update_ui = QtCore.pyqtSignal()
 
-    def __init__(self, parent: "widgets.quiz.Quiz") -> None:
+    def __init__(
+        self,
+        parent: QtWidgets.QWidget,
+        token: str,
+        quiz: m.Quiz | None = None,
+        pixmap: QtGui.QPixmap | None = None,
+        image_updated: QtCore.pyqtBoundSignal | None = None,
+    ) -> None:
         super().__init__(parent)
         self.ui = gui.dialogs.editor.Ui_QuizEditor()
         self.ui.setupUi(self)
         self.context = parent
-        self.quiz = deepcopy(self.context.quiz)
+        self.token = token
 
-        self.ui.image.setPixmap(self.context.ui.image.pixmap())
+        if quiz is not None:
+            quiz = deepcopy(quiz)
+            self.quiz_id = quiz.id
+            self.quiz = q.Quiz(
+                label=quiz.label,
+                image_url=quiz.image_url,
+                questions=quiz.questions,
+            )
+        else:
+            self.quiz_id = None
+            self.quiz = q.Quiz(label="")
+
+        if pixmap is not None:
+            self.ui.image.setPixmap(pixmap)
+        if image_updated is not None:
+            image_updated.connect(self.ui.image.setPixmap)
+
         self.ui.labelField.setText(self.quiz.label)
         self.ui.imageURLField.setText(self.quiz.image_url)
 
@@ -30,7 +56,6 @@ class QuizEditor(QtWidgets.QDialog):
         self.ui.cancelButton.clicked.connect(self.cancel_button)
         self.ui.addFieldButton.clicked.connect(self.add_button)
         self.ui.deleteButton.clicked.connect(self.delete_button)
-        self.context.image_updated.connect(self.ui.image.setPixmap)
 
         self.update_ui.connect(self._update_ui)
         self.update_ui.emit()
@@ -45,44 +70,32 @@ class QuizEditor(QtWidgets.QDialog):
         )
 
     def cancel_button(self) -> None:
-        if self.quiz != self.context.quiz:
-            if (
-                QtWidgets.QMessageBox(
-                    QtWidgets.QMessageBox.Icon.Critical,
-                    "Confirmation",
-                    "Вы хотите отменить изменения ?",
-                    QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No,
-                ).exec()
-                == QtWidgets.QMessageBox.Yes
-            ):
-                self.close()
-        else:
+        if (
+            QtWidgets.QMessageBox(
+                QtWidgets.QMessageBox.Icon.Critical,
+                "Confirmation",
+                "Вы хотите отменить изменения ?",
+                QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No,
+            ).exec()
+            == QtWidgets.QMessageBox.Yes
+        ):
             self.close()
 
     def save_button(self) -> None:
-        assert (
-            self.context.context.token is not None
-            and self.context.context.user is not None
-        )
         self.quiz.label = self.ui.labelField.text()
         self.quiz.image_url = self.ui.imageURLField.text() or None
 
-        q.update(
-            self.context.context.token,
-            self.quiz.id,
-            q.Quiz(
-                label=self.quiz.label,
-                image_url=self.quiz.image_url,
-                questions=self.quiz.questions,
-            ),
-        )
+        if self.quiz_id is None:
+            q.add(self.token, self.quiz)
+        else:
+            quiz = q.update(self.token, self.quiz_id, self.quiz)
+            if isinstance(self.context, widgets.quiz.Quiz):
+                self.context.quiz = quiz
+                self.context.update_ui.emit()
 
-        self.context.quiz = self.quiz
-        self.context.update_ui.emit()
         self.close()
 
     def delete_button(self) -> None:
-        assert self.context.context.token
         if (
             QtWidgets.QMessageBox(
                 QtWidgets.QMessageBox.Icon.Critical,
@@ -92,9 +105,10 @@ class QuizEditor(QtWidgets.QDialog):
             ).exec()
             == QtWidgets.QMessageBox.Yes
         ):
-            q.delete(self.context.context.token, self.context.quiz.id)
-            self.context.context.quizzes.remove(self.context.quiz)
-            self.context.deleteLater()
+            if self.quiz_id is not None and isinstance(self.context, widgets.quiz.Quiz):
+                q.delete(self.token, self.quiz_id)
+                self.context.context.quizzes.remove(self.context.quiz)
+                self.context.deleteLater()
             self.close()
 
     def _update_ui(self) -> None:
